+++ CHANGELOG.md
@@ -4,6 +4,17 @@ This file documents recent notable changes to this project. The format of this
 file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Simplified `list_files` to return only file names instead of unused size and
+  modified-time data.
+
+### Removed
+
+- Removed the direct `chrono` dependency.
+
 ## [0.6.0] - 2026-04-16
 
 ### Added
@@ -121,6 +132,7 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - Initial release.
 
+[Unreleased]: https://github.com/aicers/roxy/compare/0.6.0...main
 [0.6.0]: https://github.com/aicers/roxy/compare/0.5.1...0.6.0
 [0.5.1]: https://github.com/aicers/roxy/compare/0.5.0...0.5.1
 [0.5.0]: https://github.com/aicers/roxy/compare/0.4.0...0.5.0